Pull DataCategoryInput into its own file
It is getting big, and it avoids circular dependencies
